Clean up the classes a bit, simplify
